+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## NEXT
 
+- Rename product from Firebase Data Connect to Firebase SQL Connect.
+
 ## 2.2.3
 
 - Update internal `firebase-tools` dependency to 15.13.0